// frontend/src/app/models/user.ts
export default class User {
    tip: string= "";
    korisnickoIme: string = "";
    lozinka: string= "";
    ime: string= "";
    prezime: string= "";
    telefon: string= "";
    email: string="";
    aktivan: boolean= false;
    slika: string="";
    nazivInstitucije: string="";
    adresaSedista: string="";
    MB: string="";
    PIB: string="";
}